requirements.txt
pydm
git+https://github.com/githubuser0xFFFF/Qt-Advanced-Docking-System
ophyd
qtpy
typhos
fuzzywuzzy
pyyaml
